# swagger/swagger.yaml
basePath: /rest/api/1
consumes:
- application/json
definitions:
  APIKey:
    properties:
      api_key_id:
        description: APIKeyID is a generated id of the key.
        type: string
        x-go-name: APIKeyID
      api_key_secret:
        description: APIKeySecret is a generated secret, aka, the key.
        type: string
        x-go-name: APIKeySecret
      create_at:
        description: CreateAt defines when this key was created.
        example: time.Now()
        format: date-time
        type: string
        x-go-name: CreateAt
      id:
        description: ID of the key. This is auto-generated.
        format: int64
        type: integer
        x-go-name: ID
      name:
        description: Name of the key
        type: string
        x-go-name: Name
      ttl:
        description: TTL defines how long this key can live in duration.
        example: 1h10m10s
        type: string
        x-go-name: TTL
      user_id:
        description: UserID is the ID of the user to which this key belongs.
        format: int64
        type: integer
        x-go-name: UserID
    required:
    - id
    - name
    - user_id
    - api_key_id
    - api_key_secret
    - ttl
    - create_at
    title: APIKey is an api key pair generated by the user to access the api without
      the frontend.
    type: object
    x-go-package: github.com/krok-o/krok/pkg/models
...
  CommandRun:
    description: |-
      CommandRun is a single run of a command belonging to an event
      including things like, state, event, and created at.
    properties:
      command_name:
        description: CommandName is the name of the command that is being executed.
        type: string
        x-go-name: CommandName
      create_at:
        description: CreatedAt is the time when this command run was created.
        format: date-time
        type: string
        x-go-name: CreateAt
      event_id:
        description: EventID is the ID of the event that this run belongs to.
        format: int64
        type: integer
        x-go-name: EventID
      id:
        description: ID is a generatd identifier.
        format: int64
        type: integer
        x-go-name: ID
      outcome:
        description: Outcome is any output of the command. Stdout and stderr combined.
        type: string
        x-go-name: Outcome
      status:
        description: Status is the current state of the command run.
        example: running, failed, success
        type: string
        x-go-name: Status
    required:
    - id
    - event_id
    - command_name
    - status
    - create_at
    type: object
    x-go-package: github.com/krok-o/krok/pkg/models
...
  Event:
    description: |-
      Event contains details about a platform event, such as
      the repository it belongs to and the event that created it...
    properties:
      command_runs:
        description: CommandRuns contains a list of CommandRuns which executed for
          this event.
        items:
          $ref: '#/definitions/CommandRun'
        type: array
        x-go-name: CommandRuns
      create_at:
        description: CreatedAt contains the timestamp when this event occurred.
        format: date-time
        type: string
        x-go-name: CreateAt
      event_id:
        description: |-
          EvenID is the ID of the corresponding event on the given platform. If that cannot be determined
          an ID is generated.
        type: string
        x-go-name: EventID
      id:
        description: ID is a generated ID.
        format: int64
        type: integer
        x-go-name: ID
      payload:
        description: Payload defines the information received from the platform for
          this event.
        type: string
        x-go-name: Payload
      repository_id:
        description: RepositoryID contains the ID of the repository for which this
          event occurred.
        format: int64
        type: integer
        x-go-name: RepositoryID
      vcs:
        description: VCS is the ID of the platform on which this even occurred.
        format: int64
        type: integer
        x-go-name: VCS
    required:
    - id
    - event_id
    - create_at
    - repository_id
    - payload
    - vcs
    type: object
    x-go-package: github.com/krok-o/krok/pkg/models
...
  User:
    properties:
      api_keys:
        description: APIKeys contains generated api access keys for this user.
        items:
          $ref: '#/definitions/APIKey'
        type: array
        x-go-name: APIKeys
      display_name:
        description: DisplayName is the name of the user.
        type: string
        x-go-name: DisplayName
      email:
        description: Email of the user.
        type: string
        x-go-name: Email
      id:
        description: ID of the user. This is auto-generated.
        format: int64
        type: integer
        x-go-name: ID
      last_login:
        description: LastLogin contains the timestamp of the last successful login
          of this user.
        format: date-time
        type: string
        x-go-name: LastLogin
    required:
    - email
    - id
    - last_login
    title: User is a user in the Krok system.
    type: object
    x-go-package: github.com/krok-o/krok/pkg/models
...
host: krok-o.io
info:
  description: Documentation the Krok API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Krok.
  version: 0.0.1
paths:
...
  /event/{id}:
    get:
      operationId: getEvent
      parameters:
      - description: The ID of the event to retrieve
        format: int
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/Event'
        "400":
          description: invalid event id
          schema:
            $ref: '#/responses/Message'
        "404":
          description: event not found
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to get event
          schema:
            $ref: '#/responses/Message'
      summary: Get a specific event.
  /events/{repoid}:
    post:
      operationId: listEvents
      parameters:
      - description: The ID of the repository to list events for.
        format: int
        in: path
        name: repoid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
          schema:
            items:
              $ref: '#/definitions/Event'
            type: array
        "400":
          description: invalid repository id
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to list events
          schema:
            $ref: '#/responses/Message'
      summary: List events for a repository.
...
  /user:
    post:
      consumes:
      - application/json
      description: Creates a new user
      operationId: createUser
      parameters:
      - in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/User'
      produces:
      - application/json
      responses:
        "200":
          description: the created user
          schema:
            $ref: '#/definitions/User'
        "400":
          description: invalid json payload
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to create user or generating a new api key
          schema:
            $ref: '#/responses/Message'
  /user/{id}:
    delete:
      operationId: deleteUser
      parameters:
      - description: The ID of the user to delete
        format: int
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK in case the deletion was successful
        "400":
          description: in case of missing user context or invalid ID
          schema:
            $ref: '#/responses/Message'
        "404":
          description: in case of user not found
          schema:
            $ref: '#/responses/Message'
        "500":
          description: when the deletion operation failed
          schema:
            $ref: '#/responses/Message'
      summary: Deletes the given user.
    get:
      operationId: getUser
      parameters:
      - format: int
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/User'
        "400":
          description: invalid user id
          schema:
            $ref: '#/responses/Message'
        "404":
          description: user not found
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to get user
          schema:
            $ref: '#/responses/Message'
      summary: Gets the user with the corresponding ID.
...
  /user/update:
    post:
      consumes:
      - application/json
      operationId: updateUser
      parameters:
      - in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/User'
      produces:
      - application/json
      responses:
        "200":
          description: user successfully updated
          schema:
            $ref: '#/definitions/User'
        "400":
          description: invalid json payload
          schema:
            $ref: '#/responses/Message'
        "404":
          description: user not found
          schema:
            $ref: '#/responses/Message'
        "500":
          description: failed to update user
          schema:
            $ref: '#/responses/Message'
      summary: Updates an existing user.
  /users:
    post:
      description: List users
      operationId: listUsers
      produces:
      - application/json
      responses:
        "200":
          description: ""
          schema:
            items:
              $ref: '#/definitions/User'
            type: array
        "500":
          description: failed to list user
          schema:
            $ref: '#/responses/Message'
...
produces:
- application/json
responses:
  Message:
    description: Message represents an error message.
...
schemes:
- http
swagger: "2.0"